@@ -60,22 +60,14 @@ module.exports = {
         react: {
           singleton: true,
           requiredVersion: deps.react,
-          eager: true,
         },
         'react-dom': {
           singleton: true,
           requiredVersion: deps['react-dom'],
-          eager: true,
         },
-        '@emotion/core': {
+        '@emotion/react': {
           singleton: true,
-          requiredVersion: deps['react-dom'],
-          eager: true,
-        },
-        '@emotion/core': {
-          singleton: true,
-          requiredVersion: deps['react-dom'],
-          eager: true,
+          requiredVersion: deps['@emotion/react'],
         },
       },
     }),
