@@ -1,11 +1,19 @@
 import { Box } from '@mui/material'
 import React from 'react'
-import PostSliderHorizontal from 'cards/PostSliderHorizontal'
+
+import { useNavigate } from 'react-router-dom'
+
+const PostSliderHorizontal = React.lazy(() =>
+  import('cards/PostSliderHorizontal')
+)
 
 function Home() {
+  const navigate = useNavigate()
   return (
     <Box>
-      <PostSliderHorizontal />
+      <React.Suspense fallback={<div />}>
+        <PostSliderHorizontal nav={navigate} />
+      </React.Suspense>
     </Box>
   )
 }
